LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/container-inner-at-rules-expected.txt


PASS @keyframes is defined regardless of evaluation
FAIL @property is defined regardless of evaluation assert_equals: expected "20px" but got "1em"
PASS @layer order respected regardless of evaluation
PASS @font-face is defined regardless of evaluation
PASS @media works inside @container
PASS @supports works inside @container
